xapi-client-demo/requirements.txt
xapi-python~=0.1.7